requirements.txt
pandas
biopython
networkx
matplotlib
igraph
plotly
